覆盖

+++ up.sh
@@ -1,3 +1,7 @@
+git pull
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 pnpm docs:build
 rm ./docs -rf
 mv ./src/.vuepress/dist ./docs -f
